python/form_verifier/requirements.txt
cryptography==2.9.2
python-dateutil==2.8.1
requests==2.23.0
pika==1.1.0
